exit_worker/requirements.txt
kiteconnect
pymongo
dnspython
websocket-client
redis
Flask